big ol' refactor
